global-identity-automation/requirements.txt
selenium==3.141.0
pytest==5.4.3
pytest-xdist==1.27.0
pytest-rerunfailures==9.1.1
nose==1.3.7

allure-pytest==2.8.36
allure-python-commons==2.8.36
apipkg==1.5
appdirs==1.4.4
appnope==0.1.2
asgiref==3.3.1
attrs==20.3.0
backcall==0.2.0
blinker==1.4
cachetools==4.2.1
certifi==2020.11.8
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.7
cv2-tools==2.4.0
decorator==4.4.2
delayed-assert==0.3.2
diffimg==0.2.3
dill==0.3.3
distlib==0.3.1
Django==3.1.4
djangorestframework==3.12.2
entrypoints==0.3
et-xmlfile==1.0.1
execnet==1.7.1
filelock==3.0.12
flake8==3.7.9
gmail==0.6.3
google-api-core==1.26.0
google-api-python-client==2.0.2
google-auth==1.26.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
h11==0.12.0
h2==4.0.0
hpack==4.0.0
httplib2==0.19.0
humanfriendly==9.1
hyperframe==6.0.0
idna==2.10
ImageHash==4.2.0
imgcompare==2.0.1
importlib-metadata==3.1.1
iniconfig==1.1.1
ipython==7.20.0
ipython-genutils==0.2.0

jdcal==1.4.1
jedi==0.18.0
jproperties==2.1.0
kaitaistruct==0.9
mccabe==0.6.1
more-itertools==8.6.0
noompy==1.10
nose==1.3.7
numpy==1.19.4
oauth2client==4.1.3
oauthlib==3.1.0
opencv-contrib-python==4.5.1.48
opencv-python==4.4.0.46
openpyxl==3.0.5
packaging==20.7
pandas==1.1.4
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
pip==20.3
pipenv==2020.11.15
pluggy==0.13.1
progressbar==2.5
prompt-toolkit==3.0.16
protobuf==3.14.0
ptyprocess==0.7.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyGetWindow==0.0.9
Pygments==2.7.4
PyMsgBox==1.0.9

pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
PySocks==1.7.1
pytest==5.4.3
pytest-check==0.3.9
pytest-forked==1.3.0
pytest-reportportal==5.0.6
pytest-rerunfailures==9.1.1
pytest-xdist==1.27.0
python-constraint==1.4.0
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3
reportportal-client==5.0.6
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.7
rubicon-objc==0.4.0
scipy==1.6.2
secure-smtplib==0.1.1
selenium==3.141.0
selenium-update==0.0.0
selenium-wire==4.2.2
setuptools==53.0.0
shovel==0.4.0
six==1.15.0
snapyrest==0.1.1
sqlparse==0.4.1
SSIM-PIL==1.0.12
systest==5.12.1
termcolor==1.1.0
toml==0.10.2
traitlets==5.0.5
uritemplate==3.0.1
uritemplate.py==3.0.2
urllib3==1.26.3
virtualenv==20.2.1
virtualenv-clone==0.5.4
wcwidth==0.2.5
wsproto==1.0.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0
